added fixes
