--- src/components/Navbar.css
@@ -1,149 +1,0 @@
-/* 
-  1) Use a rounder font like Poppins instead of Inter or Space Grotesk.
-     Adjust weights as needed (400, 600, 700, etc.). 
-*/
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap');
-
-/* 
-  2) Navbar Container 
-     - Slightly rounded bottom corners
-     - White background
-     - Subtle box shadow for depth 
-*/
-.navbar {
-    font-family: 'Poppins', sans-serif;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    max-width: 1300px;
-    background: #ffffff;  /* was transparent, now white */
-    border-radius: 0 0 16px 16px; /* round the bottom corners */
-    box-shadow: 0 2px 8px rgba(0,0,0,0.05); 
-    padding: 0.5rem 3%;
-    z-index: 1000;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: 0 auto;
-    transition: background 0.3s ease;
-}
-
-/*
-  🔹 Navbar Links Container 
-  Flex layout with spacing 
-*/
-.nav-links {
-    display: flex;
-    list-style: none;
-    gap: 1.5rem;
-    padding: 0;
-    margin: 0;
-}
-
-/*
-  🔹 Each nav item 
-     - Use position: relative if you want the underline effect 
-*/
-.nav-links li {
-    position: relative;
-    display: flex;
-    align-items: center;
-}
-
-/*
-  🔹 Link Styling
-     - Slightly round hover background
-     - No uppercase for a softer vibe, or remove if you prefer uppercase 
-*/
-.nav-links li a {
-    text-decoration: none;
-    color: #205EA3;                 /* Default color: brand blue */
-    font-size: 1rem;
-    font-weight: 500;               /* Poppins 500 or 600 works well */
-    /* text-transform: uppercase; */ /* Remove or comment out for a rounder, friendlier look */
-    padding: 0.75rem 1rem;          /* Slightly larger for a bigger clickable area */
-    cursor: pointer;
-    display: inline-block;
-    position: relative;
-    transition: all 0.2s ease-in-out;
-    border-radius: 8px;             /* Rounded corners on links themselves */
-}
-
-/*
-  🔹 If you want to keep the underline effect, 
-     here's a subtle version with a slight horizontal expansion 
-*/
-.nav-links li a::after {
-    content: '';
-    display: block;
-    width: 0;
-    height: 2px;
-    background: #205EA3; 
-    transition: width 0.3s ease-in-out;
-    position: absolute;
-    bottom: -5px;
-    left: 0;
-    border-radius: 2px;
-}
-
-/*
-  🔹 Hover / Focus 
-     - Light scale up
-     - Darker shade
-     - Reveal underline 
-     - Optional background for a pill-like hover 
-*/
-.nav-links li a:hover {
-    transform: scale(1.08);
-    color: #003366;             /* Darker shade on hover */
-    background: transparent;        /* Gentle background highlight on hover */
-}
-
-.nav-links li a:hover::after {
-    width: 100%;
-    background: #003366;
-}
-
-/* 
-  🔹 Responsive Adjustments 
-*/
-@media (max-width: 1024px) {
-    .navbar {
-        max-width: 100%;
-        padding: 0.5rem 5%;
-        border-radius: 0; /* Possibly remove round corners at narrower widths */
-    }
-    .nav-links {
-        gap: 1rem;
-    }
-}
-
-@media (max-width: 768px) {
-    .navbar {
-        flex-direction: column;
-        height: auto;
-        padding: 1rem 5%;
-    }
-    .nav-links {
-        flex-direction: column;
-        gap: 0.8rem;
-        text-align: center;
-        padding-top: 0.5rem;
-    }
-}
-
-@media (max-width: 480px) {
-    .navbar {
-        text-align: center;
-    }
-    .nav-links {
-        flex-direction: column;
-        gap: 0.5rem;
-        padding-top: 0.5rem;
-    }
-    .nav-links li a {
-        font-size: 0.95rem;
-        padding: 0.6rem 0.8rem;
-    }
-}
