+++ src/components/benefits/benefits.css
@@ -154,10 +154,22 @@
 
 /* Responsive */
 @media (max-width: 768px) {
+  .timeline-steps {
+    flex-direction: column;
+    align-items: center;
+    overflow-x: hidden;
+  }
+
   .timeline-step {
+    min-width: unset;
+    width: 100%;
+    text-align: center;
+  }
+  
+  /* .timeline-step {
     min-width: 120px;
     padding: 0.8rem;
-  }
+  } */
   .timeline h2 {
     font-size: 1.2rem;
   }
